validado para não incluir notificações caso paciente foi a obito
